FIX(#13): FIX VOLUNTEER MODEL
Fixed Volunteer model by adding it to the Person app and deleting Volunteer app. Also, Pillow 9.4.0 was added to requirements
