+++ person/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
-from .models import GodFather, ASEMUser, Worker
+from .models import GodFather, ASEMUser, Worker, Volunteer
 
 admin.site.register(GodFather)
 admin.site.register(ASEMUser)
 admin.site.register(Worker)
+admin.site.register(Volunteer)
