+++ requirements.txt
@@ -2,3 +2,4 @@ Django==4.1.4
 python-decouple==3.6
 mysqlclient==2.1.1
 coverage==7.2.1
+pillow==9.4.0
