--- volunteer/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from .models import Volunteer
-
-# Register your models here.
-admin.site.register(Volunteer)
-
